separate files for classes: Fruit, Item, Alcohol
