+++ shm/src/alcohol.cpp
@@ -1,0 +1,4 @@
+#include "alcohol.hpp"
+
+Alcohol::Alcohol(std::string name, size_t amount, double basePrice)
+    : Cargo(name, amount, basePrice) {}
