+++ shm/src/fruit.cpp
@@ -1,0 +1,4 @@
+#include "fruit.hpp"
+
+Fruit::Fruit(std::string name, size_t amount, double basePrice)
+    : Cargo(name, amount, basePrice) {}
